refactor(spice): split chunk_executor_actor into per-shard/coordinator/tracker files (6/n)
